Update class-wc-rest-product-reviews-controller.php
When migrating reviews it would be practical to also set the verified owner status already in the request. Otherwise, this has to be done manually in teh database.
It would also be good to add this in the endpoint of the REST service, as well as the review creation date. This would make migrations easier.